requirements.txt
mss
pyperclip
Pillow
pytesseract